create Table wrapper element to conditionally render Table or TreeGridTable
