+++ packages/@react-spectrum/table/stories/Table.stories.tsx
@@ -152,7 +152,7 @@ export const Static: TableStory = {
   name: 'static'
 };
 
-let columns = [
+export let columns = [
   {name: 'Foo', key: 'foo'},
   {name: 'Bar', key: 'bar'},
   {name: 'Baz', key: 'baz'}
@@ -311,96 +311,6 @@ export const DynamicSelectedKeys: TableStory = {
       exclude: /selectionMode/
     }
   }
-};
-
-// TODO: add stories with variable expandedKeys (will need to wait for )
-export const StaticExpandableRows = {
-  args: {
-    'aria-label': 'TableView with static expandable rows',
-    width: 500,
-    height: 200
-  },
-  render: (args) => (
-    <TableView {...args} expandedKeys={['test']}>
-      <TableHeader>
-        <Column key="foo">Foo</Column>
-        <Column key="bar">Bar</Column>
-        <Column key="baz">Baz</Column>
-      </TableHeader>
-      <TableBody>
-        <Row key="test">
-          <Cell>Lvl 1 Foo 1</Cell>
-          <Cell>Lvl 1 Bar 1</Cell>
-          <Cell>Lvl 1 Baz 1</Cell>
-          <Row>
-            <Cell>Lvl 2 Foo 1</Cell>
-            <Cell>Lvl 2 Bar 1</Cell>
-            <Cell>Lvl 2 Baz 1</Cell>
-            <Row>
-              <Cell>Lvl 3 Foo 1</Cell>
-              <Cell>Lvl 3 Bar 1</Cell>
-              <Cell>Lvl 3 Baz 1</Cell>
-            </Row>
-          </Row>
-          <Row>
-            <Cell>Lvl 2 Foo 2</Cell>
-            <Cell>Lvl 2 Bar 2</Cell>
-            <Cell>Lvl 2 Baz 2</Cell>
-          </Row>
-        </Row>
-      </TableBody>
-    </TableView>
-  ),
-  name: 'static with expandable rows'
-};
-
-let nestedItems = [
-  {test: 'Lvl 1 Test 1', foo: 'Lvl 1 Foo 1', bar: 'Lvl 1 Bar 1', yay: 'Lvl 1 Yay 1', baz: 'Lvl 1 Baz 1', childRows: [
-    {test: 'Lvl 2 Test 1', foo: 'Lvl 2 Foo 1', bar: 'Lvl 2 Bar 1', yay: 'Lvl 2 Yay 1', baz: 'Lvl 2 Baz 1', childRows: [
-      {test: 'Lvl 3 Test 1', foo: 'Lvl 3 Foo 1', bar: 'Lvl 3 Bar 1', yay: 'Lvl 3 Yay 1', baz: 'Lvl 3 Baz 1'}
-    ]},
-    {test: 'Lvl 2 Test 2', foo: 'Lvl 2 Foo 2', bar: 'Lvl 2 Bar 2', yay: 'Lvl 2 Yay 2', baz: 'Lvl 2 Baz 2'}
-  ]}
-];
-
-function DynamicExpandableRows(props: SpectrumTableProps<unknown>) {
-  let [expandedKeys, setExpandedKeys] = useState(new Set([]));
-
-  return (
-    <Flex direction="column">
-      <ActionButton onPress={() => setExpandedKeys('all')}>Expand all</ActionButton>
-      <ActionButton onPress={() => setExpandedKeys(new Set([]))}>Collapse all</ActionButton>
-      <ActionButton onPress={() => setExpandedKeys(new Set(['Lvl 1 Foo 1']))}>Expand subset</ActionButton>
-      <TableView {...props} expandedKeys={expandedKeys} onExpandedChange={action('onExpandedChange')}>
-        <TableHeader columns={columns}>
-          {column => <Column>{column.name}</Column>}
-        </TableHeader>
-        <TableBody items={nestedItems}>
-          {item =>
-            (<Row key={item.foo} childItems={item.childRows}>
-              {(key) => {
-                // Note: The "item" here will reflect the child Row's values from nestedItems
-                return <Cell>{item[key]}</Cell>;
-              }}
-            </Row>)
-          }
-        </TableBody>
-      </TableView>
-    </Flex>
-
-  );
-}
-
-export const DyanmicExpandableRows = {
-  args: {
-    'aria-label': 'TableView with dynamic expandable rows',
-    width: 500,
-    height: 200
-  },
-  render: (args) => (
-    <DynamicExpandableRows {...args} />
-  ),
-  name: 'dynamic with expandable rows'
 };
 
 export const StaticNestedColumns: TableStory = {
